Notification: int8 datatype is replaced by bigint

+++ go/src/socialapi/workers/notification/sql/004-table.sql
@@ -3,9 +3,9 @@
 -- ----------------------------
 DROP TABLE IF EXISTS "notification"."notification";
 CREATE TABLE "notification"."notification" (
-    "id" int8 NOT NULL DEFAULT nextval('notification.notification_id_seq'::regclass),
-    "account_id" int8 NOT NULL,
-    "notification_content_id" int8 NOT NULL,
+    "id" bigint NOT NULL DEFAULT nextval('notification.notification_id_seq'::regclass),
+    "account_id" bigint NOT NULL,
+    "notification_content_id" bigint NOT NULL,
     "glanced" bool NOT NULL,
     "subscribed_at" timestamp(6) WITH TIME ZONE,
     "unsubscribed_at" timestamp(6) WITH TIME ZONE,
@@ -28,8 +28,8 @@ ALTER TYPE "notification"."notification_content_type_constant_enum" OWNER TO "so
 
 DROP TABLE IF EXISTS "notification"."notification_content";
 CREATE TABLE "notification"."notification_content" (
-    "id" int8 NOT NULL DEFAULT nextval('notification.notification_content_id_seq'::regclass),
-    "target_id" int8 NOT NULL,
+    "id" bigint NOT NULL DEFAULT nextval('notification.notification_content_id_seq'::regclass),
+    "target_id" bigint NOT NULL,
     "type_constant" "notification"."notification_content_type_constant_enum",
     "created_at" timestamp(6) WITH TIME ZONE NOT NULL DEFAULT now()
 )
@@ -41,9 +41,9 @@ GRANT SELECT, INSERT ON "notification"."notification_content" TO "socialapplicat
 -- ----------------------------
 DROP TABLE IF EXISTS "notification"."notification_activity";
 CREATE TABLE "notification"."notification_activity" (
-    "id" int8 NOT NULL DEFAULT nextval('notification.notification_activity_id_seq'::regclass),
-    "actor_id" int8 NOT NULL,
-    "notification_content_id" int8 NOT NULL,
+    "id" bigint NOT NULL DEFAULT nextval('notification.notification_activity_id_seq'::regclass),
+    "actor_id" bigint NOT NULL,
+    "notification_content_id" bigint NOT NULL,
     "obsolete" bool NOT NULL,
     "created_at" timestamp(6) WITH TIME ZONE NOT NULL DEFAULT now()
 )
